move for split
